Update mypy to latest stable release
Upgrade mypy from 1.15.0 to 1.20.0 to keep the project's static analysis tooling current on Python 3.12. Adjust the small set of code paths surfaced by stricter type checking around pandas timestamp coercion, news filter TypedDict writes, toast message guards, async task handling, trade widget/controller typing, and QApplication access so the full plutus_terminal package type-checks cleanly without changing intended runtime behavior.

+++ plutus_terminal/ui/widgets/perps_trade.py
@@ -675,11 +675,11 @@ class MarketTradeWidget(QtWidgets.QWidget):
         """Get amount."""
         return self.amount_box.value()
 
-    def get_take_profit(self) -> Decimal:
+    def get_take_profit(self) -> float:
         """Get take profit."""
         return self.take_profit_box.value()
 
-    def get_stop_loss(self) -> Decimal:
+    def get_stop_loss(self) -> float:
         """Get stop loss."""
         return self.stop_loss_box.value()
 
@@ -778,11 +778,11 @@ class LimitTradeWidget(QtWidgets.QWidget):
         """Get price."""
         return self.target_price_box.value()
 
-    def get_take_profit(self) -> Decimal:
+    def get_take_profit(self) -> float:
         """Get take profit."""
         return self.take_profit_box.value()
 
-    def get_stop_loss(self) -> Decimal:
+    def get_stop_loss(self) -> float:
         """Get stop loss."""
         return self.stop_loss_box.value()
